+++ src/main/resources/mybatis/mappers/RequestMapper.xml
@@ -3,7 +3,7 @@
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 
-<mapper namespace="com.woosan.hr_system.report.dao.RequestDAO">
+<mapper namespace="request">
 <!-- ==================================================생성 SQL문=======================================================-->
     <!-- 요청 생성 -->
     <insert id="createRequest" parameterType="map">
